Moved fixture and unit test + revisit Unit test
